Fixed computer reference problems using the methods: "unlinkOldReferences", to unlink previously references, "saveNewOneToOneReferences", to save processor reference and "saveNewManyToOneReferences", to save ram memories and storage devices; Unnecessary methods have been removed; Code standardization

+++ src/main/java/com/pml/dto/ComputerDTO.java
@@ -31,8 +31,8 @@ public class ComputerDTO implements Serializable {
 	private Integer operatingSystem;
 	private Integer operatingSystemArchitecture;
 	private boolean onTheDomain;
-	private Integer totalRamMemory;
-	private Integer totalStorageMemory;
+	private Double totalRamMemory;
+	private Double totalStorageMemory;
 	
 	public ComputerDTO() {
 	}
@@ -207,19 +207,19 @@ public class ComputerDTO implements Serializable {
 		this.onTheDomain = onTheDomain;
 	}
 
-	public Integer getTotalRamMemory() {
+	public Double getTotalRamMemory() {
 		return totalRamMemory;
 	}
 
-	public void setTotalRamMemory(Integer totalRamMemory) {
+	public void setTotalRamMemory(Double totalRamMemory) {
 		this.totalRamMemory = totalRamMemory;
 	}
 	
-	public Integer getTotalStorageMemory() {
+	public Double getTotalStorageMemory() {
 		return totalStorageMemory;
 	}
 
-	public void setTotalStorageMemory(Integer totalStorageMemory) {
+	public void setTotalStorageMemory(Double totalStorageMemory) {
 		this.totalStorageMemory = totalStorageMemory;
 	}
