+++ src/main/java/com/pml/dto/StorageDeviceNewDTO.java
@@ -16,7 +16,7 @@ public class StorageDeviceNewDTO implements Serializable {
 	private String description;
 	private boolean itWorks = true;
 	@NotNull(message = "This field is mandatory")
-	private Integer sizeInGB;
+	private Double sizeInGB;
 	@NotNull(message = "This field is mandatory")
 	private Integer architecture;
 	@NotNull(message = "This field is mandatory")
@@ -58,11 +58,11 @@ public class StorageDeviceNewDTO implements Serializable {
 		this.itWorks = itWorks;
 	}
 	
-	public Integer getSizeInGB() {
+	public Double getSizeInGB() {
 		return sizeInGB;
 	}
 
-	public void setSizeInGB(Integer sizeInGB) {
+	public void setSizeInGB(Double sizeInGB) {
 		this.sizeInGB = sizeInGB;
 	}
